pyproject.toml
[project]
name = "final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pyarrow>=23.0.0",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "transformers>=4.57.6",
]
